pyproject.toml
[project]
name = "ffpy"
version = "0.1.0"
description = "Fantasy Football Point Projection Web App"
readme = "README.md"
authors = [
    { name = "RickArko", email = "rick.arko17@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "polars>=1.36.1",
    "duckdb>=1.4.3",
]

[project.scripts]
ffpy = "ffpy.app:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "coverage>=7.13.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
analysis = [
    "jupyter>=1.0.0",
    "matplotlib~=3.10",
    "seaborn>=0.13.0",
    "numpy>=1.26.0",
]
